+++ file_backup.php
@@ -5,6 +5,10 @@ require_once __DIR__ . "/config.php";
 require_once __DIR__.'/Drive.php';
 use w3lifer\Google\Drive;
 try {    
+    if (!FILE_BACKUP) {
+        logs("File Backup is disabled");
+        exit(1);
+    }
     $folder_excluded = FOLDER_EXCLUDE; 
     $destination_filename = ZIP_FILE_NAME . "_" . date("Y-m-d") . ".zip";
     $destination = BACKUP_PATH . "/" . $destination_filename;
@@ -16,7 +20,7 @@ try {
     } else {
         logs($response);
     }
-    if ($hasBackup) {
+    if ($hasBackup && EXPORT_FILE_TO_GOOGLE_DRIVE) {
         if (file_exists($destination)) {
             logs("Files Backup Created : ". $destination_filename . "(". getFileSize($destination) . ")", "INFO");
             if (empty(GOOGLE_SERVICE_ACCOUNT_JSON)) {
@@ -30,7 +34,9 @@ try {
                 logs("File Upload To Google Drive Failed");
             } else {
                 logs("File Uploaded : $fileId", "INFO");
-                unlink($destination);
+                if(LOCAL_DELETE_FILE_AFTER_EXPORT_TO_GOOGLE_DRIVE){
+                    unlink($destination);
+                }
             }
         } else {
             logs("$destination not found");
